feat: diet plan update 로직 구현

+++ src/main/java/com/ssafy/yamyam_coach/repository/daily_diet/MyBatisDailyDietRepository.java
@@ -60,6 +60,10 @@ public class MyBatisDailyDietRepository implements DailyDietRepository {
 
     @Override
     public int deleteByDietPlanAndDateInBatch(Long dietPlanId, List<LocalDate> datesToDelete) {
+        if (datesToDelete == null || datesToDelete.isEmpty()) {
+            return 0;
+        }
+        
         return dailyDietMapper.deleteByDietPlanAndDateInBatch(dietPlanId, datesToDelete);
     }
